Move gammapy.script.cta_irf to gammapy.irf.io
