+++ gammapy/irf/__init__.py
@@ -12,4 +12,5 @@ from .psf_analytical import *
 from .psf_king import *
 from .energy_dispersion import *
 from .irf_stack import *
+from .io import *
