+++ gammapy/scripts/__init__.py
@@ -3,5 +3,4 @@
 Gammapy command line interface (scripts).
 """
 from .spectrum_pipe import *
-from .cta_irf import *
 from .cta_sensitivity import *
